requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
cryptography==42.0.2
python-dotenv==1.0.0
PyJWT==2.8.0

# Testing Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
httpx==0.25.2

# Development Dependencies
black==23.10.1
flake8==6.1.0
mypy==1.7.1

# Security Dependencies
bandit==1.7.5

# Development Tools
ipython==8.17.2
jupyter==1.0.0

# Monitoring & Telemetry
prometheus-client==0.19.0
psutil==5.9.6
opentelemetry-sdk==1.21.0

# Dependency Resolution
packaging>=22.0 